fixed bugs

+++ app/src/main/java/com/hod/finalapp/model/firebase/StorageManager.java
@@ -285,7 +285,10 @@ public class StorageManager
         ArrayList<String> itemsPictureUrls = iItem.getPicturesUrls();
         for(String url : itemsPictureUrls)
         {
+            if(!url.equals(""))
+            {
             mFirebaseStorage.getReferenceFromUrl(url).delete();
+            }
         }
     }
